+++ lib/src/lib.rs
@@ -1,6 +1,7 @@
 //#![warn(missing_docs)]
 
 pub mod common;
+pub mod entities;
 pub mod offline;
 pub mod online;
 
@@ -9,8 +10,8 @@ mod android;
 
 pub use common::context::*;
 pub use common::error::*;
-pub use common::json::*;
 pub use common::*;
+pub use entities::*;
 
 // Re-exports
 pub use bitcoin;
@@ -21,4 +22,4 @@ pub use serde_json;
 pub use structopt;
 
 pub type Result<R> = std::result::Result<R, Error>;
-pub type PSBT = bitcoin::util::psbt::PartiallySignedTransaction;
+pub type BitcoinPSBT = bitcoin::util::psbt::PartiallySignedTransaction;
