+++ lib/src/offline/random.rs
@@ -14,10 +14,10 @@ pub struct RandomOptions {
 }
 
 impl OfflineContext {
-    pub fn create_key(&self, opt: &RandomOptions) -> Result<MasterSecretJson> {
+    pub fn create_key(&self, opt: &RandomOptions) -> Result<MasterSecret> {
         let sec = rand::thread_rng().gen::<[u8; 32]>();
         let xpriv = ExtendedPrivKey::new_master(self.network, &sec)?;
-        let master_key = MasterSecretJson::new(self.network, xpriv, &opt.key_name)?;
+        let master_key = MasterSecret::new(self.network, xpriv, &opt.key_name)?;
         self.write_keys(&master_key)?;
 
         Ok(master_key)
@@ -30,7 +30,6 @@ mod tests {
     use crate::common::tests::rnd_string;
     use crate::offline::random::RandomOptions;
     use miniscript::descriptor::DescriptorSecretKey;
-    use miniscript::DescriptorPublicKey;
     use std::str::FromStr;
 
     impl RandomOptions {
@@ -57,13 +56,13 @@ mod tests {
     #[test]
     fn test_descriptor_key() {
         let xpub = "[a15f432e/48'/1'/0']tpubDDoLq7YG6qr18Paph1uJ8F2ncVuSh2DjkixS6CX37nCiJusecE82JXvFmfZh8hp86Bm7sv7Pkprv5phMXn1r49TU6YrDidGmemFAL1PNWXi/0/*";
-        assert!(DescriptorPublicKey::from_str(&xpub).is_ok());
+        assert!(miniscript::DescriptorPublicKey::from_str(&xpub).is_ok());
 
         let xpub = "tpubDDoLq7YG6qr18Paph1uJ8F2ncVuSh2DjkixS6CX37nCiJusecE82JXvFmfZh8hp86Bm7sv7Pkprv5phMXn1r49TU6YrDidGmemFAL1PNWXi";
-        assert!(DescriptorPublicKey::from_str(&xpub).is_ok());
+        assert!(miniscript::DescriptorPublicKey::from_str(&xpub).is_ok());
 
         let xpub = "[a15f432e/48'/1'/0']tpubDDoLq7YG6qr18Paph1uJ8F2ncVuSh2DjkixS6CX37nCiJusecE82JXvFmfZh8hp86Bm7sv7Pkprv5phMXn1r49TU6YrDidGmemFAL1PNWXi";
-        assert!(DescriptorPublicKey::from_str(&xpub).is_ok());
+        assert!(miniscript::DescriptorPublicKey::from_str(&xpub).is_ok());
 
         assert!(DescriptorSecretKey::from_str("xprv9s21ZrQH143K3yGb6gtghzHH4MPaEHGPN48sxoyYd4EdrQcaSVP2dxZS2vRwoKny1KRS5xMMyGunA3WkToah7ZmJ2fFtGK8vBBBiBkVFmTM").is_ok());
     }
